reorganise rheology part
